+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.2.0
+
+- Fixed a bug that flashed signed off content before loading.
+- Changed the default behavior of Google sign-in to prompt for account:
+
+```dart
+abstract class FirebaseAuth {
+  Future<FirebaseUser> googleSignIn({bool prompt: true});
+}
+```
+
 ## 0.1.1
 
 - Removed a `print` statement that was always occurring.